+++ packages/assets-controllers/src/MultichainAssetsController/MultichainAssetsController.test.ts
@@ -773,4 +773,49 @@ describe('MultichainAssetsController', () => {
       });
     });
   });
+
+  describe('getAssetMetadata', () => {
+    it('returns the metadata for a given asset', async () => {
+      const { messenger } = setupController({
+        state: {
+          accountsAssets: {
+            [mockSolanaAccount.id]: mockHandleRequestOnAssetsLookupReturnValue,
+          },
+          assetsMetadata: mockGetMetadataReturnValue.assets,
+        } as MultichainAssetsControllerState,
+      });
+
+      const assetId = 'solana:EtWTRABZaYq6iMfeYKouRu166VU2xqa1/slip44:501';
+
+      const metadata = messenger.call(
+        'MultichainAssetsController:getAssetMetadata',
+        assetId,
+      );
+
+      expect(metadata).toStrictEqual(
+        mockGetMetadataReturnValue.assets[assetId],
+      );
+    });
+
+    it('returns undefined if the asset metadata is not found', async () => {
+      const { messenger } = setupController({
+        state: {
+          accountsAssets: {
+            [mockSolanaAccount.id]: mockHandleRequestOnAssetsLookupReturnValue,
+          },
+          assetsMetadata: mockGetMetadataReturnValue.assets,
+        } as MultichainAssetsControllerState,
+      });
+
+      const assetId =
+        'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/token:EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v';
+
+      const metadata = messenger.call(
+        'MultichainAssetsController:getAssetMetadata',
+        assetId,
+      );
+
+      expect(metadata).toBeUndefined();
+    });
+  });
 });