+++ packages/bridge-controller/src/utils/bridge.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable n/no-process-env */
+import { Contract } from '@ethersproject/contracts';
 import { abiERC20 } from '@metamask/metamask-eth-abis';
 import type { Hex } from '@metamask/utils';
-import { Contract } from 'ethers';
 
 import {
   getEthUsdtResetData,