+++ packages/profile-sync-controller/src/sdk/authentication.test.ts
@@ -1,8 +1,6 @@
-import {
-  MOCK_ACCESS_JWT,
-  MOCK_SRP_LOGIN_RESPONSE,
-  arrangeAuthAPIs,
-} from './__fixtures__/mock-auth';
+import type { Eip1193Provider } from 'ethers';
+
+import { arrangeAuthAPIs } from './__fixtures__/auth';
 import type { MockVariable } from './__fixtures__/test-utils';
 import { arrangeAuth, arrangeMockProvider } from './__fixtures__/test-utils';
 import { JwtBearerAuth } from './authentication';
@@ -14,6 +12,7 @@ import {
   UnsupportedAuthTypeError,
   ValidationError,
 } from './errors';
+import { MOCK_ACCESS_JWT, MOCK_SRP_LOGIN_RESPONSE } from './mocks/auth';
 import * as Eip6963MetamaskProvider from './utils/eip-6963-metamask-provider';
 import { Env, Platform } from '../shared/env';
 
@@ -576,13 +575,14 @@ describe('Authentication - SRP Default Flow - signMessage() & getIdentifier()', 
 
     // Sign Message
     await expect(auth.signMessage('not formatted message')).rejects.toThrow(
-      ValidationError,
+      'Message must start with "metamask:"',
     );
   });
 
   it('successfully uses default SRP flow', async () => {
     arrangeAuthAPIs();
     const { auth } = arrangeAuth('SRP', MOCK_SRP, { signing: undefined });
+
     arrangeProvider();
 
     const accessToken = await auth.getAccessToken();
@@ -593,6 +593,29 @@ describe('Authentication - SRP Default Flow - signMessage() & getIdentifier()', 
 
     const message = await auth.signMessage('metamask:test message');
     expect(message).toBeDefined();
+  });
+});
+
+describe('Authentication - rejects when calling unrelated methods', () => {
+  it('rejects when calling SRP methods in SiWE flow', async () => {
+    const { auth } = arrangeAuth('SiWE', MOCK_ADDRESS);
+
+    expect(() => auth.setCustomProvider({} as Eip1193Provider)).toThrow(
+      UnsupportedAuthTypeError,
+    );
+  });
+
+  it('rejects when calling SiWE methods in SRP flow', async () => {
+    const { auth } = arrangeAuth('SRP', MOCK_SRP);
+
+    expect(() =>
+      auth.prepare({
+        address: MOCK_ADDRESS,
+        chainId: 1,
+        domain: 'https://metamask.io',
+        signMessage: async () => 'MOCK_SIWE_SIGNATURE',
+      }),
+    ).toThrow(UnsupportedAuthTypeError);
   });
 });